feat(document-loader): add web loader and improve directory loader
- Integrate a new WebLoader for handling web documents.
- Improve the accuracy of the DirectoryLoader's `support` method by handling exceptions.
- Optimize document chunking process in KnowledgeService when adding documents to vector store.
- Correct invocation of methods in ConversationalContextualCompressionRetrievalQAChain and ConversationalFastRetrievalQAChain.
- Update text loader to support markdown files and make various other improvements.
- Refactor and enhance the functionality of the Command plugin related to document handling.

BREAKING CHANGE: The addition of WebLoader and changes to the DirectoryLoader may affect existing document loading processes.

+++ src/plugins/command.ts
@@ -28,7 +28,12 @@ export async function apply(
             fallback: false
         })
         .action(async ({ options, session }, path) => {
+            if (path.startsWith('http')) {
+                return `目前暂不支持 url：${path}`
+            }
+
             path = await copyDocument(ctx, path, options.copy)
+
             const loader = ctx.chatluna_knowledge.loader
 
             const supported = await loader.support(path)
@@ -149,18 +154,20 @@ export async function setRoomKnowledgeConfig(
 }
 
 async function deleteDocument(ctx: Context, filePath: string, db: string) {
-    try {
-        await fs.access(filePath)
-    } catch (e) {
-        filePath = path.resolve(
-            ctx.baseDir,
-            'data/chathub/knowledge/data',
-            filePath
-        )
-    }
+    if (!filePath.startsWith('http')) {
+        try {
+            await fs.access(filePath)
+        } catch (e) {
+            filePath = path.resolve(
+                ctx.baseDir,
+                'data/chathub/knowledge/data',
+                filePath
+            )
+        }
 
-    if (filePath.startsWith(ctx.baseDir)) {
-        await fs.rm(filePath, { recursive: true })
+        if (filePath.startsWith(ctx.baseDir)) {
+            await fs.rm(filePath, { recursive: true })
+        }
     }
 
     await ctx.chatluna_knowledge.deleteDocument(filePath, db)
